requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
bitarray==2.8.3
bitstring==4.1.2
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
decouple==0.0.7
deprecation==2.1.0
Django==4.2.6
django-elasticsearch-dsl==8.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.14.0
ecdsa==0.18.0
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
esptool==4.6.2
Faker==22.0.0
Flask==3.0.0
fonttools==4.46.0
functions-framework==3.5.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.2
Pillow==10.1.0
promise==2.3
psycopg2==2.9.9
pycparser==2.21
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
reedsolo==1.7.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
text-unidecode==1.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1